django/requirements.txt
django
djangorestframework
django-model-utils
dj-rest-auth
drf-yasg